Fix problem with hardcoded path in virtualenv script, reduce spammy output, fix incorrect make target
Normally, when creating a virtualenv, the script at bin/activate
contains a full, hardcoded path to the location of the virtualenv. This
does not work well when the virtualenv is created within the container
and we want to `source` it outside the container.

Update the `_deploy` target to not have `package/package.zip` as a
dependency, because it has no way to satisfy all the dependencies
required to create it - it needs to be invoked in a different container.